refactor(bus): invert error coupling + split Knative ingress (P4b/4)
- bus/error.rs no longer names HandlerError: the From<HandlerError> conversion
  and classification move to microsvc (HandlerError::transport_error_kind);
  From<RepositoryError> for TransportError moves to the outbox bridge (which
  knows both types). Bus core's error.rs is now microsvc-free.
- Knative split: the Service-coupled HTTP ingress (cloud_events_router /
  ingress_handler / CloudEvent parsing) moves to microsvc::knative_ingress;
  bus/knative.rs keeps only the Message/SubscriptionPlan-only manifest helpers
  (knative_triggers, sanitize_k8s_name). Two test crates updated.

229 lib tests green; cargo check --all-features clean.

Implements [[tasks/bus-decomposition-phase1]]
