+++ src/bus/error.rs
@@ -9,8 +9,6 @@
 use std::error::Error;
 use std::fmt;
 
-use crate::microsvc::HandlerError;
-
 /// Whether a [`TransportError`] should be retried.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, serde::Serialize, serde::Deserialize)]
 pub enum TransportErrorKind {
@@ -103,40 +101,6 @@ impl TransportError {
     pub fn message(&self) -> &str {
         &self.message
     }
-
-    /// Classify a [`HandlerError`] for transport retry purposes.
-    ///
-    /// Transient failures are retryable: repository errors, not-found results,
-    /// and otherwise-unclassified errors. Not-found is transient in an
-    /// at-least-once event-driven system because an out-of-order delivery can
-    /// reference an aggregate or projection that has not been created yet, and a
-    /// later redelivery resolves the race. (This also keeps a handler-raised
-    /// [`HandlerError::NotFound`] consistent with a repository-raised
-    /// [`RepositoryError::NotFound`](crate::RepositoryError), which arrives
-    /// wrapped in `Repository(_)`.)
-    ///
-    /// Deterministic failures — unknown routing, payload decode failures,
-    /// business rejections, authorization, and guard rejections — are permanent
-    /// because redelivering the identical message cannot change the outcome. An
-    /// entity that is genuinely missing forever still drains to the failure
-    /// policy once the transport's retry ceiling is reached, so defaulting
-    /// not-found to retryable does not risk an unbounded loop.
-    ///
-    /// This is a default. Handlers that know better can return a classification
-    /// directly, and the runner's [`FailurePolicy`](super::FailurePolicy) can
-    /// override what happens to a permanent error.
-    pub fn classify_handler_error(error: &HandlerError) -> TransportErrorKind {
-        match error {
-            HandlerError::Repository(_) | HandlerError::NotFound(_) | HandlerError::Other(_) => {
-                TransportErrorKind::Retryable
-            }
-            HandlerError::UnknownCommand(_)
-            | HandlerError::DecodeFailed(_)
-            | HandlerError::Rejected(_)
-            | HandlerError::Unauthorized(_)
-            | HandlerError::GuardRejected(_) => TransportErrorKind::Permanent,
-        }
-    }
 }
 
 impl fmt::Display for TransportError {
@@ -154,30 +118,6 @@ impl Error for TransportError {
         self.source
             .as_ref()
             .map(|source| source.as_ref() as &(dyn Error + 'static))
-    }
-}
-
-impl From<HandlerError> for TransportError {
-    fn from(error: HandlerError) -> Self {
-        let kind = Self::classify_handler_error(&error);
-        Self {
-            kind,
-            message: error.to_string(),
-            source: Some(Box::new(error)),
-        }
-    }
-}
-
-impl From<crate::RepositoryError> for TransportError {
-    /// Repository/store failures (lock contention, storage hiccups, stale-claim
-    /// conflicts) are treated as retryable: they are usually transient, and a
-    /// later redelivery or re-claim resolves them.
-    fn from(error: crate::RepositoryError) -> Self {
-        Self {
-            kind: TransportErrorKind::Retryable,
-            message: error.to_string(),
-            source: Some(Box::new(error)),
-        }
     }
 }
 
@@ -218,51 +158,4 @@ mod tests {
         assert_eq!(err.message(), "publish timed out");
     }
 
-    #[test]
-    fn transient_handler_errors_are_retryable() {
-        use crate::repository::RepositoryError;
-
-        for error in [
-            HandlerError::Repository(RepositoryError::NotFound { id: "agg-1".into() }),
-            // A handler-raised not-found is transient too: it is usually an
-            // out-of-order delivery race, and stays consistent with the
-            // repository-raised not-found above.
-            HandlerError::NotFound("agg-1".into()),
-            HandlerError::Other("boom".into()),
-        ] {
-            assert_eq!(
-                TransportError::classify_handler_error(&error),
-                TransportErrorKind::Retryable,
-                "{error} should be retryable"
-            );
-        }
-    }
-
-    #[test]
-    fn deterministic_handler_errors_are_permanent() {
-        for error in [
-            HandlerError::UnknownCommand("x".into()),
-            HandlerError::DecodeFailed("x".into()),
-            HandlerError::Rejected("x".into()),
-            HandlerError::Unauthorized("x".into()),
-            HandlerError::GuardRejected("x".into()),
-        ] {
-            assert_eq!(
-                TransportError::classify_handler_error(&error),
-                TransportErrorKind::Permanent,
-                "{error} should be permanent"
-            );
-        }
-    }
-
-    #[test]
-    fn from_handler_error_preserves_classification_and_source() {
-        let err: TransportError = HandlerError::Rejected("invalid".into()).into();
-        assert!(err.is_permanent());
-        assert!(err.source().is_some());
-
-        let err: TransportError =
-            HandlerError::Other(Box::<dyn Error + Send + Sync>::from("infra")).into();
-        assert!(err.is_retryable());
-    }
 }
